Add a full settings page shared with the popup (fixes #7).
Extract shared preference rendering and persistence so popup and options
page write to the same storage, and link to the settings page from the popup.
