+++ template-extension/manifest.json
@@ -49,7 +49,8 @@
     "storage"
   ],
   "options_ui": {
-    "page": "options.html"
+    "page": "options.html",
+    "open_in_tab": true
   },
   "browser_specific_settings": {
     "gecko": {
