+++ template-extension/options.html
@@ -1,14 +1,13 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
   <head>
     <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <title>Tab Sorter</title>
+    <link rel="icon" type="image/png" href="icons/icon-48.png" />
+    <link rel="stylesheet" href="tab-sorter.css" />
   </head>
-
-  Please use the extension's popup to configure your preferences.
-
-  <img
-    src="icons/icon-96.png"
-    alt="Tab Sorter's Icon"
-    style="max-width: 96px; width: 96px"
-  />
+  <body>
+    <script type="module" src="options.js"></script>
+  </body>
 </html>
